requirements.txt
# requirements.txt cho FastAPI, Prometheus, Jaeger, test, lint
fastapi
uvicorn
prometheus_fastapi_instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
pytest
requests
flake8
black
isort
